:construction: Add option

+++ API/src/main/java/fr/maxlego08/essentials/api/vault/VaultItem.java
@@ -13,7 +13,7 @@ public interface VaultItem {
 
     ItemStack getDisplayItemStack(ComponentMessage adventureComponent);
 
-    void addQuantity(int amount);
+    void addQuantity(long amount);
 
     void removeQuantity(long amount);
 }
